fix feign client conflict

+++ ProductService/src/main/java/com/mazen/ProductService/service/ProductService.java
@@ -1,10 +1,12 @@
 package com.mazen.ProductService.service;
+import com.mazen.ProductService.dto.OrderStatus;
 import com.mazen.ProductService.dto.ProductDetailsResponse;
 import com.mazen.ProductService.dto.ProductResponse;
 import com.mazen.ProductService.dto.request.post.ProductRequest;
 import com.mazen.ProductService.dto.request.update.ProductUpdate;
 import com.mazen.ProductService.exceptions.BadRequestException;
 import com.mazen.ProductService.exceptions.NotFoundException;
+import com.mazen.ProductService.kafka.ProductProducer;
 import com.mazen.ProductService.model.Product;
 import com.mazen.ProductService.model.ProductCategory;
 import com.mazen.ProductService.model.ProductImage;
@@ -13,6 +15,7 @@ import com.mazen.ProductService.repository.ProductCategoryRepository;
 import com.mazen.ProductService.repository.ProductSpecsRepository;
 import com.mazen.ProductService.repository.ProductRepository;
 import com.mazen.ProductService.service.feignClient.FileServiceClient;
+import com.mazen.ProductService.service.feignClient.OrderServiceClient;
 import com.mazen.ProductService.service.feignClient.SaleServiceClient;
 import com.mazen.ProductService.util.PagedResponse;
 import feign.FeignException;
@@ -40,32 +43,38 @@ public class ProductService {
     private final ModelMapper modelMapper;
     private final MappingService mappingService;
     private final FileServiceClient fileServiceClient;
-    private final SaleServiceClient saleServiceClient;
+    private final ProductProducer productProducer;
+    private final OrderServiceClient orderServiceClient;
 
     @Transactional
     public void createProduct(ProductRequest productRequest,String authorization) {
-        log.info(authorization);
 
+        log.info("Product request {}",productRequest.toString());
+        // check if product category exists in DB
         ProductCategory productCategory = productCategoryRepository.findByCategory
                 (productRequest.getProductCategory()).orElseThrow(()->
                 new NotFoundException("Not found product category with the name"));
 
+        // build product object
         Product product = Product.builder()
                 .title(productRequest.getTitle())
                 .price(productRequest.getPrice())
                 .productCategory(productCategory)
                 .build();
 
+        // build specs list of product
         List<ProductSpecs> productSpecs = productRequest
                 .getProductSpecsRequests()
                 .stream()
                 .map(productSpecsRequest -> mappingService
                         .convertToProductSpecs(productSpecsRequest,product))
                 .toList();
-        product.setProductSpecs(productSpecs);
 
+        // set the values
+        product.setProductSpecs(productSpecs);
         Product product1 = productRepository.save(product);
 
+        // loop over images from request and sent to file storage service
         productRequest
                 .getProductImageRequests().forEach(productImageRequest -> {
                     try {
@@ -77,15 +86,13 @@ public class ProductService {
     }
 
     @Transactional
-    public void deleteProduct(String id,String authorization){
-        try{
-            if(!fileServiceClient.deletePhotos(id,authorization).getBody()){
-                throw new BadRequestException("Failed to delete images of product");
-            }
+    public void deleteProduct(String id){
+        List<Object> orders = orderServiceClient.getOrdersByProductId(id,List.of(OrderStatus.CURRENT
+                ,OrderStatus.DELIVERED,OrderStatus.PACKING,OrderStatus.SHIPPED));
+        if(!orders.isEmpty()) {
+            throw new BadRequestException("There are some orders currently...");
         }
-        catch (Exception ex){
-            throw new BadRequestException(ex.getMessage());
-        }
+        productProducer.sendProductId(id);
         productRepository.deleteById(id);
     }
 
@@ -167,7 +174,7 @@ public class ProductService {
         productRepository.save(product);
     }
 
-    public List<ProductResponse> getProductsByIds(List<String> ids){
+    public List<ProductResponse> getProductsByIds(List<String> ids,String authorization){
         List<Product> products = productRepository.findAllById(ids);
         // Extract IDs of found products
         List<String> foundIds = products.stream()
@@ -182,126 +189,41 @@ public class ProductService {
         if (!missingIds.isEmpty()) {
             throw new NotFoundException("Products not found with ids: " + missingIds);
         }
-        return products.stream().map(mappingService::createProductResponse).toList();
+        return products.stream().map(product->mappingService.createProductResponse(product,authorization)).toList();
     }
 
-//    @Transactional
-//    public void changeItemInventory(String productId, Size size, Colors color, int numberOfItems){
-//        Product product = getProductByIdOptional(productId);
-//        Optional<ProductColor> productColor =
-//                productColorRepository.findByProductSizeAndColor(size,color,product);
-//        if(productColor.isPresent()){
-//            productColor.get().setNumberInStock(numberOfItems);
-//            productColorRepository.save(productColor.get());
-//            return;
-//        }
-//        Optional<ProductColor> productColor1 =
-//                productColorRepository.findByProductByColor(color,product);
-//        if(productColor1.isPresent()){
-//            productColor1.get().setSize(size);
-//            productColor1.get().setColors(color);
-//            productColorRepository.save(productColor1.get());
-//            return;
-//        }
-//        ProductColor productColor2 = ProductColor.builder()
-//                .colors(color)
-//                .size(size)
-//                .product(product)
-//                .numberInStock(numberOfItems)
-//                .build();
-//
-//        product.getProductColors().add(productColor2);
-//        productRepository.save(product);
-//    }
-//
-//    @Transactional
-//    public void changeItemInventoryOnlyColor(String productId, Colors color, int numberOfItems){
-//        Product product = getProductByIdOptional(productId);
-//
-//        Optional<ProductColor> productColor1 =
-//                productColorRepository.findByProductByColor(color,product);
-//        if(productColor1.isPresent()){
-//            productColor1.get().setColors(color);
-//            productColorRepository.save(productColor1.get());
-//            return;
-//        }
-//        ProductColor productColor2 = ProductColor.builder()
-//                .colors(color)
-//                .product(product)
-//                .numberInStock(numberOfItems)
-//                .build();
-//
-//        product.getProductColors().add(productColor2);
-//        productRepository.save(product);
-//    }
-    public boolean isProductExist(String product_id){
-        Optional<Product> product = productRepository.findById(product_id);
-        return product.isPresent();
-    }
-
-    public List<String> isProductsExists(List<String> ids){
-        List<String> idsNotFound = new ArrayList<>();
-        log.info(ids.toString());
-        ids.forEach(id->{
-            if(!isProductExist(id)){
-                idsNotFound.add(id);
-            }
-        });
-        return idsNotFound;
-    }
-
-    public ProductDetailsResponse getProductDetailsById(String id){
+    public ProductDetailsResponse getProductDetailsById(String id,String authorization){
         Product product = getProductByIdOptional(id);
-        return mappingService.createProductDetailsResponse(product);
+        return mappingService.createProductDetailsResponse(product,authorization);
     }
 
-    public PagedResponse<ProductResponse> getAllProductByCategory(ProductCategory category, int page, int size){
+    public PagedResponse<ProductResponse> getAllProductByCategory(ProductCategory category,
+                                                                  int page, int size,
+                                                                  String authorization){
         Pageable pageable = PageRequest.of(page,size, Sort.by("createdAt").descending());
         Page<Product> products= productRepository.getAllProductsByCategory(pageable, category);
         List<ProductResponse> productResponses = products
-                .map(mappingService::createProductResponse).toList();
+                .map(product -> mappingService.createProductResponse(product,authorization)).toList();
         return new PagedResponse<>(productResponses,
                 page,size,products.getTotalElements(),
                 products.getTotalPages(),products.isLast());
     }
 
-    public PagedResponse<ProductResponse> getAllProductsRandom(int page , int size){
+    public PagedResponse<ProductResponse> getAllProductsRandom(int page , int size,String authorization){
         Pageable pageable = PageRequest.of(page,size, Sort.by("createdAt").descending());
         Page<Product> products= productRepository.getAllProductsRandom(pageable);
         List<ProductResponse> productResponses = products
-                .map(mappingService::createProductResponse).toList();
+                .map(product -> mappingService.createProductResponse(product,authorization)).toList();
         return new PagedResponse<>(productResponses,
                 page,size,products.getTotalElements(),
                 products.getTotalPages(),products.isLast());
     }
 
-    public double getPriceOfAllProducts(List<String> ids){
-        double sum = 0.0;
-        List<Integer> discounts = new ArrayList<>();
-
-        try{
-             discounts = saleServiceClient.getProductsDiscountByIds(ids);
-        }
-        catch (FeignException ex){
-            log.error(ex.getLocalizedMessage());
-        }
-
-        int index = 0;
-        for(String id : ids){
-            Product product = getProductByIdOptional(id);
-            int discount = discounts.get(index);
-            double price =  (product.getPrice()*discount)/100.0;
-            sum+= price;
-            index++;
-        }
-         return sum;
-    }
-
-    public PagedResponse<ProductResponse> getAllProduct(int page, int size) {
+    public PagedResponse<ProductResponse> getAllProduct(int page, int size,String authorization) {
         Pageable pageable = PageRequest.of(page,size, Sort.by("createdAt").descending());
         Page<Product> products= productRepository.findAll(pageable);
         List<ProductResponse> productResponses = products
-                .map(mappingService::createProductResponse).toList();
+                .map(product -> mappingService.createProductResponse(product,authorization)).toList();
         return new PagedResponse<>(productResponses,
                 page,size,products.getTotalElements(),
                 products.getTotalPages(),products.isLast());
